+++ src/main/java/com/example/demo/repositories/ProductRepository.java
@@ -1,12 +1,10 @@
 package com.example.demo.repositories;
 
-
 import org.springframework.data.repository.CrudRepository;
 
 import com.example.demo.entities.Product;
 
-public interface ProductRepository extends CrudRepository<Product,Integer>
-{
-	 Product findByName(String name);
+public interface ProductRepository extends CrudRepository<Product, Long> {
+	Product findByName(String name);
 
 }
